refactor: 토큰 로직 Header -> Providers로 더 최상단 이동
- 의미없는 스타일 삭제
