+++ src/app/layout.tsx
@@ -1,37 +1,39 @@
+'use client';
+
 import React from 'react';
-import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import '@/styles/globals.css';
 import Script from 'next/script';
+import { usePathname } from 'next/navigation';
 import Header from './components/Header';
 import Footer from './components/Footer';
 import Providers from './providers';
 
 const inter = Inter({ subsets: ['latin'] });
 
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+const RootLayout = ({ children }: { children: React.ReactNode }) => {
+  const pathname = usePathname();
+
+  return (
+    <html lang="en" className="light" suppressHydrationWarning>
+      <head>
+        <Script
+          strategy="beforeInteractive"
+          src={`//dapi.kakao.com/v2/maps/sdk.js?appkey=${process.env.NEXT_PUBLIC_MAP_KEY}&autoload=false&libraries=services`}
+        />
+      </head>
+      <body className={`${inter.className}`}>
+        <Providers>
+          {pathname.includes('details') ||
+          pathname.includes('new') ||
+          pathname.includes('mypage') ? null : (
+            <Header />
+          )}
+          <div className="pb-[48px] pt-[61px]">{children}</div>
+          {pathname.includes('chatroom') ? null : <Footer />}
+        </Providers>
+      </body>
+    </html>
+  );
 };
-
-const RootLayout = ({ children }: { children: React.ReactNode }) => (
-  <html lang="en" className="light" suppressHydrationWarning>
-    <head>
-      <Script
-        strategy="beforeInteractive"
-        src={`//dapi.kakao.com/v2/maps/sdk.js?appkey=${process.env.NEXT_PUBLIC_MAP_KEY}&autoload=false&libraries=services`}
-      />
-    </head>
-    <body className={`${inter.className} flex min-h-screen flex-col`}>
-      <div className="flex-grow">
-        <Header />
-        <div className="pb-[48px] pt-[61px]">
-          <Providers>{children}</Providers>
-        </div>
-        <Footer />
-      </div>
-    </body>
-  </html>
-);
-
 export default RootLayout;
