+++ src/app/page.tsx
@@ -1,5 +1,11 @@
 import React from 'react';
+import type { Metadata } from 'next';
 import Image from 'next/image';
+
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+};
 
 const Home = () => (
   <div>
